feat: changes utils to validators. Adds a json validation class
